Consolidate on a single source of team colours
