+++ api/web/src/base/cot-icon.spec.ts
@@ -119,12 +119,7 @@ describe('COT.as_rendered', () => {
         expect(rendered.properties!.icon).toBeUndefined();
     });
 
-    // Regression: marker-stroke-color was set by styleProperties() and read by
-    // the skittle layer's circle-stroke-color expression in styles.ts, but was
-    // missing from RENDERED_PROPERTIES - the whitelist as_rendered() uses to
-    // slim a Feature down before it reaches the vector tile. It was silently
-    // stripped here, so every team's marker border fell back to the style's
-    // default white regardless of the computed value.
+    // marker-stroke-color must survive as_rendered()
     it('carries marker-stroke-color through to the rendered Feature', () => {
         const rendered = COT.as_rendered({
             id: 'render-contact-stroke',
