+++ api/web/src/lib/geolocate/main.ts
@@ -1,5 +1,6 @@
 import { Marker, LngLat } from 'maplibre-gl';
 import type { Map as MapLibreMap, IControl, ControlPosition } from 'maplibre-gl';
+import { TEAM_COLORS, strokeColorFor } from '../../utils/team-colors.ts';
 
 export type GeolocateControlOptions = {
     /**
@@ -39,48 +40,6 @@ const defaultOptions: ResolvedOptions = {
 const STYLE_ELEMENT_ID = 'cloudtak-geolocate-styles';
 
 const DEFAULT_PUCK_COLOR = '#1da1f2';
-
-// Mirrors the team -> marker-color mapping used for rendered CoT markers in
-// base/cot.ts (TEAM_COLORS there) so the self puck matches the rendered self
-// CoT marker and the ATAK client rendering the same team colour. These are
-// ATAK's actual RGB values (Icon2525cIconAdapter.teamToColor()), not CSS/UI
-// framework theme colours - see cot.ts for the full rationale.
-export const TEAM_COLORS: Record<string, string> = {
-    White: '#FFFFFF',
-    Yellow: '#FFFF00',
-    Orange: '#FF7700',
-    Magenta: '#FF00FF',
-    Red: '#FF0000',
-    Maroon: '#7F0000',
-    Purple: '#7F007F',
-    'Dark Blue': '#00007F',
-    Blue: '#0000FF',
-    Cyan: '#00FFFF',
-    Teal: '#007F7F',
-    Green: '#00FF00',
-    'Dark Green': '#007F00',
-    Brown: '#A0714F',
-};
-
-/**
- * Black or white border colour for a fill, chosen by relative luminance
- * (ITU-R BT.601 coefficients) so the puck stays visible against either a
- * light or dark basemap. Falls back to white (the prior fixed behaviour)
- * for a value that cannot be parsed as `#rrggbb`. Mirrors strokeColorFor()
- * in cot.ts.
- */
-export function strokeColorFor(hex: string): string {
-    const match = /^#?([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})$/i.exec(hex);
-    if (!match) return '#ffffff';
-
-    const r = parseInt(match[1], 16);
-    const g = parseInt(match[2], 16);
-    const b = parseInt(match[3], 16);
-
-    const luminance = (0.299 * r + 0.587 * g + 0.114 * b) / 255;
-
-    return luminance > 0.6 ? '#000000' : '#ffffff';
-}
 
 /**
  * A MapLibre {@link IControl} that renders a live user-location puck (dot,
@@ -315,11 +274,6 @@ export class GeolocateControl implements IControl {
     private applyColor(): void {
         if (this.dotElement) {
             this.dotElement.style.backgroundColor = this.color;
-            // A fixed white border (the CSS default, see .cloudtak-geolocate-dot
-            // below) is invisible for a light puck colour (White/Yellow/Cyan
-            // teams) against a light basemap. Switch to a black border for a
-            // light fill, mirroring the same luminance-based choice made for
-            // other users' rendered CoT markers (see strokeColorFor() in cot.ts).
             this.dotElement.style.borderColor = strokeColorFor(this.color);
         }
         if (this.headingElement) {
